Solve generate button hidden.
